apps/api/requirements.txt
Flask
psycopg2-binary
python-dotenv
SQLAlchemy
google-cloud-secret-manager
gunicorn
flask-cors
google-cloud-compute
pytest
pytest-flask
google-cloud-network-security>=0.6.0